move File component from the Files directory to its own component
