+++ app/components/file/file_info.tsx
