+++ app/components/file/file_options_icon.tsx
