+++ app/components/file/image_file.tsx
