+++ app/components/file/image_file_overlay.tsx
